+++ internal/common.go
@@ -2,7 +2,7 @@ package internal
 
 const (
 	ENV_ID_DESC                     = "The Id to the environment that this resource will be created in"
-	UNABLE_TO_CREATE_SVIX_CLIENT    = "Unable to create svix client, this error should not happen, please contact the developers"
+	UNABLE_TO_CREATE_SVIX_CLIENT    = "Unable to create Svix client"
 	REQUIRES_PRO_OR_ENTERPRISE_PLAN = "<strong>Requires Pro or Enterprise plan</strong>, "
 	REQUIRES_ENTERPRISE_PLAN        = "<strong>Requires Enterprise plan</strong>, "
 )