+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+
+* Add `svix_autoconfig` resource for [Webhooks AutoConfig](https://docs.svix.com/receiving/webhooks-autoconfig)
+* Provider `token` and `server_url` are now optional at configure time (still required for non-AutoConfig resources)